requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cloudinary==1.29.0
cycler==0.11.0
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-userforeignkey==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
fonttools==4.29.1
imageio==2.15.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tifffile==2022.2.9
urllib3==1.26.8
whitenoise==6.0.0
